+++ python/sglang/multimodal_gen/runtime/disaggregation/transport/rdma/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: Apache-2.0
+"""RDMA transport: P2P tensor transfer via Mooncake TransferEngine."""
