+++ python/sglang/multimodal_gen/runtime/disaggregation/transport/relay/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Relay transport: ZMQ-based tensor serialization through DiffusionServer."""
